+++ packages/frontend/src/pages/user/followers.vue
@@ -4,11 +4,11 @@ SPDX-License-Identifier: AGPL-3.0-only
 -->
 
 <template>
-<PageWithHeader :actions="headerActions" :tabs="headerTabs">
+<PageWithHeader v-model:tab="tab" :actions="headerActions" :tabs="headerTabs">
 	<div class="_spacer" style="--MI_SPACER-w: 1000px;">
 		<Transition name="fade" mode="out-in">
-			<div v-if="user">
-				<XFollowList :user="user" type="followers"/>
+			<div v-if="user" :key="`${user.id}:${tab}`">
+				<XFollowList :user="user" :type="tab === 'commonFollowers' ? 'commonFollowers' : 'followers'"/>
 			</div>
 			<MkError v-else-if="error" @retry="fetchUser()"/>
 			<MkLoading v-else/>
@@ -24,6 +24,7 @@ import XFollowList from './follow-list.vue';
 import { misskeyApi } from '@/utility/misskey-api.js';
 import { definePage } from '@/page.js';
 import { i18n } from '@/i18n.js';
+import { $i } from '@/i.js';
 
 const props = withDefaults(defineProps<{
 	acct: string;
@@ -32,10 +33,12 @@ const props = withDefaults(defineProps<{
 
 const user = ref<null | Misskey.entities.UserDetailed>(null);
 const error = ref<any>(null);
+const tab = ref<'all' | 'commonFollowers'>('all');
 
 function fetchUser(): void {
 	if (props.acct == null) return;
 	user.value = null;
+	error.value = null;
 	misskeyApi('users/show', Misskey.acct.parse(props.acct)).then(u => {
 		user.value = u;
 	}).catch(err => {
@@ -43,13 +46,22 @@ function fetchUser(): void {
 	});
 }
 
-watch(() => props.acct, fetchUser, {
+watch(() => props.acct, () => {
+	tab.value = 'all';
+	fetchUser();
+}, {
 	immediate: true,
 });
 
 const headerActions = computed(() => []);
 
-const headerTabs = computed(() => []);
+const headerTabs = computed(() => $i != null && user.value != null && $i.id !== user.value.id ? [{
+	key: 'all',
+	title: i18n.ts.all,
+}, {
+	key: 'commonFollowers',
+	title: i18n.ts.commonFollowers,
+}] : []);
 
 definePage(() => ({
 	title: i18n.ts.user,